Update FirstGame.cpp
The random value generator was turned into a function. There's a huge speed related issue as srand takes time in whole seconds and every random number has to wait 1 second before it is generated. It can be solved by feeding it time in nano seconds, which will be fixed later.